SSO and index.html
SSO sign in works, and we get their email. Now we also filter it and make sure it is @student.bth.se email otherwise we deny the login. Also it saves your last signed in session currently, but this could be time based until session expires.

Needed index.html in templates

Using 127.0.0.1 instead of localhost. But if you want to use localhost add it as host in app.run at the bottom of server.py.